Split CI tests into separate actions (so they don't take 2 hours)
